bugfix: this is needed, i guess it just deletes everything without it

+++ src/EntityDb.MongoDb/Documents/LeaseDocument.cs
@@ -187,6 +187,11 @@ namespace EntityDb.MongoDb.Documents
             IReadOnlyCollection<ILease> deleteLeases
         )
         {
+            if (deleteLeases.Count == 0)
+            {
+                return;
+            }
+
             var deleteLeasesQuery = new DeleteLeasesQuery(entityId, deleteLeases);
 
             await DeleteMany
